Don't just link to the external search (#2)
There are anchors for each twtxt in the generated page, and now
clicking on the link will point to the twtxt in this actual page
(instead of using an external service), but the link to the search
tool will still exist (only now marked as 'search'.

This change could probably be made prettier, but it works, for now.

Co-authored-by: Marcos Marado <mindboosternoori@gmail.com>
Reviewed-on: https://git.mills.io/yarnsocial/twtxt2html/pulls/2
Co-authored-by: Marcos Marado <marado@noreply@mills.io>
Co-committed-by: Marcos Marado <marado@noreply@mills.io>