Change markdown parser (#4)
Hi!

This PR changes the markdown parser being used, fixing #3.
It also fixes a small typo.
Finally, it adds a test for #3, but... to be honest the test should be rewrittten (it should be calling the actual FormatTwt function being tested, so a change there doesn't go unnoticed by an outdated test case), but I don't really have the energy to tackle that (and dive into lextwt, which could use some documentation O:-) ) - you might want to get it as is, fix it yourself, or ditch it.

The commit history here is a bit of a mess, since I was just trying to figure stuff out by myself - these should probably be squashed into only two commits, one for the typo fix, the other for the markdown parser change.

Co-authored-by: Marcos Marado <mindboosternoori@gmail.com>
Reviewed-on: https://git.mills.io/yarnsocial/twtxt2html/pulls/4
Co-authored-by: Marcos Marado <marado@noreply@mills.io>
Co-committed-by: Marcos Marado <marado@noreply@mills.io>